# calculator/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Controller API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /calculator/offers:
    post:
      summary: Calculate loan offers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanStatementRequestDto'
      responses:
        '200':
          description: Successful calculation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoanOfferDto'
  /calculator/calc:
    post:
      summary: Calculate credit
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScoringDataDto'
      responses:
        '200':
          description: Successful calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditDto'
components:
  schemas:
    LoanOfferDto:
      type: object
      properties:
        statementId:
          type: string
          format: uuid
        requestedAmount:
          type: number
          format: float
        totalAmount:
          type: number
          format: float
        term:
          type: integer
          format: int32
        monthlyPayment:
          type: number
          format: float
        rate:
          type: number
          format: float
        isInsuranceEnabled:
          type: boolean
        isSalaryClient:
          type: boolean
    CreditDto:
      type: object
      properties:
        amount:
          type: number
          format: float
        term:
          type: integer
          format: int32
        monthlyPayment:
          type: number
          format: float
        rate:
          type: number
          format: float
        psk:
          type: number
          format: float
        isInsuranceEnabled:
          type: boolean
        isSalaryClient:
          type: boolean
        paymentSchedule:
          type: array
          items:
            $ref: '#/components/schemas/PaymentScheduleElementDto'
    PaymentScheduleElementDto:
      type: object
      properties:
        number:
          type: integer
          format: int32
        date:
          type: string
          format: date
        totalPayment:
          type: number
          format: float
        interestPayment:
          type: number
          format: float
        debtPayment:
          type: number
          format: float
        remainingDebt:
          type: number
          format: float
    LoanStatementRequestDto:
      type: object
      properties:
        amount:
          type: number
          format: float
        term:
          type: integer
          format: int32
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        email:
          type: string
          format: email
        birthdate:
          type: string
          format: date
        passportSeries:
          type: string
        passportNumber:
          type: string
    ScoringDataDto:
      type: object
      properties:
        amount:
          type: number
          format: float
        term:
          type: integer
          format: int32
        firstName:
          type: string
        lastName:
          type: string
        middleName:
          type: string
        gender:
          type: string
          enum:
            - MALE
            - FEMALE
            - NON_BINARY
        birthdate:
          type: string
          format: date
        passportSeries:
          type: string
        passportNumber:
          type: string
        passportIssueDate:
          type: string
          format: date
        passportIssueBranch:
          type: string
        maritalStatus:
          type: string
          enum:
            - SINGLE
            - MARRIED
            - DIVORCED
        dependentAmount:
          type: integer
          format: int32
        employment:
          $ref: '#/components/schemas/EmploymentDto'
        accountNumber:
          type: string
        isInsuranceEnabled:
          type: boolean
        isSalaryClient:
          type: boolean
    EmploymentDto:
      type: object
      properties:
        employmentStatus:
          type: string
          enum:
            - UNEMPLOYED
            - EMPLOYED
            - SELF_EMPLOYED
            - BUSINESS_OWNER
        employerINN:
          type: string
        salary:
          type: number
          format: float
        position:
          type: string
          enum:
            - EMPLOYEE
            - MANAGER
            - TOP_MANAGER
        workExperienceTotal:
          type: integer
          format: int32
        workExperienceCurrent:
          type: integer
          format: int32